Add drone creation feature
A new file "AddDrone.fs" has been added that contains the logic for creating a drone in the drone API project. This includes the functions and types to create a new drone and save it in the drone context. This addition expands the drone API's capabilities to handle drone creation requests.