Improve sorting and pagination in GetDrones
Created better sorting functionality for the GetDrones function in the Drone API. This includes ordering the drones by Make and Model and improving the pagination by adding a default page size and letting the user choose the page size. Removed the Page type from the Drone domain as it is no longer needed.