Implement RabbitMQ and flight registered handler function
The messaging system RabbitMQ has been incorporated into the 'Drone.Api' program configuration. Moreover, a new function to handle the registration of flights has been inserted into 'ProcessDataAsyncHandler' in the 'Drone.Service' namespace. This function discriminates and prints messages based on whether the flight was rejected or, if registered, gives the flight path.