Simplify routing setup and comment out SharedMessage handling
The commit has updated the RabbitMQ configuration for Drone.Service and Drone.Api to use conventional routing without specific queue and exchange names. Also, the handler for SharedMessage in NotifySubscribersHandler has been commented out as it's stealing the message that the service should process. I need to look into processing a message in two different locations. These changes are part of the effort to streamline the management of messages within our services.