Enable Newtonsoft for serialization
The Newtonsoft library has been enabled for serialization in the Drone.Service and Drone.Api programs. This change allows discriminated unions to be used as messages.

The message handlers in the api project have been commented out because they hijack the message in the api project, thus preventing it from being processed in the service project.